add comfirm

+++ src/components/Event.tsx
@@ -7,22 +7,25 @@ import { RootState } from "../domain/entity/RootState";
 import eventActions from "../store/Events/action";
 
 import { Button } from "@material-ui/core";
-import { TableRow, TableCell } from "@material-ui/core";
+import { TableRow, TableCell, Checkbox } from "@material-ui/core";
 
 const Event: React.FC = () => {
-  const events = useSelector((state: RootState) => state.event);
+  const events = useSelector((state: RootState) => state.events);
   const dispatch = useDispatch();
   const handleDelete = (id: number) => {
-    dispatch(eventActions.deleteEvent(id));
+    const result = window.confirm("イベントを本当に削除しても良いですか？");
+    if (result) {
+      dispatch(eventActions.deleteEvent(id));
+    }
   };
   return (
     <>
       {events.map(
         (event: IEvent, i) =>
           event.id !== 0 && (
-            <TableRow key={i}>
+            <TableRow key={i} hover>
               <TableCell>
-                <input type="checkbox" />
+                <Checkbox />
               </TableCell>
               <TableCell>{event.id}</TableCell>
               <TableCell>{event.title}</TableCell>
